Basic Tower Range, Tower Data, Tower Enemy Detection, Player Health

+++ Assets/Scripts/Scriptable Objects/Enemy Data/EnemyData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "BlankEnemyData", menuName = "Scriptable Objects/EnemyData")]
 public class EnemyData : ScriptableObject
 {
-    public float EnemyDamage;
+    public int EnemyDamage;
     public float EnemyHealth;
     public float EnemySpeed;
 }
